randstruct: Create basis for unit test module
Build the unit tests with (in the build directory):

  $ make ASTTests

Run the unit tests with (in the build directory):

  $ ./tools/clang/unittests/AST/ASTTests --gtest_filter=StructureLayoutRandomization*

This initial commit outlines the specification for Clang Randstruct.
Unit tests must exercise each path of a new chunk of code for
Randstruct. Among these unit tests are also special cases outlined in
the RFC feedback on Phabricator: https://reviews.llvm.org/D59254#1429401

Signed-off-by: Connor Kuehl <cipkuehl@gmail.com>